Created library skeleton
Since the project requires also a userspace library to test the module, I created the skeleton
for its source. It comprehends
- the library itself, compiled with ar in lib/ folder
- a fiber executable for testing the library

I also created a Makefile that deals with the binary and the library compiling task. Just hit
make lib or make bin, by default make bin is executed is make is called with no parameters.

Module was moved to /module folder and I added some references from which I was inpired by.

P.S.: I also added task to VSCode for the library, so as in the module they can be accessed with
ctrl+shift+p -> Run Build Task.
P.S.2: I added Clean Dmesg task to module project to clean the entire dmesg and better understanding
the module output

Signed-off-by: Gabriele Proietti Mattia <gabry.gabry@hotmail.it>