Implemented library safe exit, module exit_fibered() and auto fd close
The module offers now another service `exit_fibered` that is called by the
library in the safe_cleanup function and therefore when a fiber accidentally
ends. The service just frees all the structures kmallocated by the module.

Another important point related to exit_fibered was fixed: when a process calls
SwitchToFiber since the module service forces the process to change the eip,
the device that was opened by the library for calling ioctl is not closed and
therefore any other calls to `open` fails causing the library to not execute
any other ioctl command (as `exit_fibered`, for example). For solving this
problem I added an utility function in the *device* part of the module that is
called `close_device_descriptor`. As documented, the function simply scrolls
on all the file descriptors opened by the process and when it finds a descriptor
associated to a device with the same minor number of our fiber device, it
closes it by calling `sys_close()`.

Signed-off-by: Gabriele Proietti Mattia <gabry.gabry@hotmail.it>