FlsGetValue must return a long, so a structure is needed for returning it
Since ioctl can return only int, when the benchmark tried to assinging long
a 4 bytes were loss, causing segmentation fault since the benchmark program
uses Set and Get value for assigning pointers (that are 8bytes long in x64).

Benchmark suite is still not working but this first issue is solved.